Update validity tests
